requirements.txt
beautifulsoup4==4.14.3
fastapi==0.115.12
jinja2==3.1.6
lxml==6.0.4
numpy==2.0.2
pandas==2.3.3
python-dotenv==1.1.1
requests==2.32.5
uvicorn==0.34.2
yfinance==0.2.66
